+++ src/math.rs
@@ -1,44 +1,44 @@
 /// Returns a 32-bit float if the value is negative (-1.0), positive (1.0), or zero (0.0)
-pub fn signf(value: f64) -> f64 {
+pub fn signf(value: f32) -> f32 {
 	if value > 0.0 { 1.0 } else if value < 0.0 { -1.0 } else { 0.0 }
 }
 
 /// Returns a 32-bit integer if the value is negative (-1), positive (1), or zero (0)
-pub fn signi(value: i64) -> i64 {
+pub fn signi(value: i32) -> i32 {
 	if value > 0 { 1 } else if value < 0 { -1 } else { 0 }
 }
 
 /// Returns a linearly interpolated 32-bit float between a and b, using t as a percentage of... 'betweenness'?
-pub fn lerpf(a: f64, b: f64, t: f64) -> f64 {
+pub fn lerpf(a: f32, b: f32, t: f32) -> f32 {
 	a + (b - a) * t
 }
 
 /// Returns a linearly interpolated 32-bit integer between a and b, using t as a percentage of... 'betweenness'?
-pub fn lerpi(a: i64, b: i64, t: f64) -> i64 {
-	((a + (b - a)) as f64 * t).floor() as i64
+pub fn lerpi(a: i32, b: i32, t: f32) -> i32 {
+	((a + (b - a)) as f32 * t).floor() as i32
 }
 
-pub fn unlerpf(value: f64, min: f64, max: f64) -> f64 {
+pub fn unlerpf(value: f32, min: f32, max: f32) -> f32 {
 	(value - min) / (max - min)
 }
 
 /// Returns a vector rotated in 2D space.
-pub fn rotate2(x: f64, y: f64, a: f64) -> (f64, f64) {
+pub fn rotate2(x: f32, y: f32, a: f32) -> (f32, f32) {
 	let (acos, asin) = (a.cos(), a.sin());
 	let ax = (x * acos) - (y * asin);
 	let ay = (x * asin) + (y * acos);
 	return (ax, ay);
 }
 
-pub fn dot2(x0: f64, y0: f64, x1: f64, y1: f64) -> f64 {
+pub fn dot2(x0: f32, y0: f32, x1: f32, y1: f32) -> f32 {
 	x0 * x1 + y0 * y1
 }
 
-pub fn cross2(x0: f64, y0: f64, x1: f64, y1: f64) -> f64 {
+pub fn cross2(x0: f32, y0: f32, x1: f32, y1: f32) -> f32 {
 	(x0 * y1) - (y0 * x1)
 }
 
-pub fn barycentric2(v1x: f64, v1y: f64, v2x: f64, v2y: f64, v3x: f64, v3y: f64) -> (f64, f64, f64) {
+pub fn barycentric2(v1x: f32, v1y: f32, v2x: f32, v2y: f32, v3x: f32, v3y: f32) -> (f32, f32, f32) {
 	let b0 = (v2x - v1x, v2y - v1y);
 	let b1 = (v3x - v1x, v3y - v1y);
 	let b2 = (v1x - v2x, v1y - v2y);
